+++ go.mod
@@ -3,6 +3,8 @@ module trace
 go 1.14
 
 require (
+	github.com/aws/aws-sdk-go v1.29.15
+	github.com/cheekybits/genny v1.0.0
 	github.com/gin-contrib/static v0.0.0-20200916080430-d45d9a37d28e
 	github.com/gin-gonic/gin v1.6.3
 	github.com/sirupsen/logrus v1.4.2
