+++ pkg/database/event.go
@@ -2,8 +2,6 @@ package database
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -39,103 +37,6 @@ type Event struct {
 	Source    EventSource        `json:"source"`
 }
 
-// EventRef is a reference to an event which, when serialized, will return
-// the json of the referenced event.
-//
-// Be careful for circular references.
-type EventRef primitive.ObjectID
-
-func (ref EventRef) GetBSON() (interface{}, error) {
-	return primitive.ObjectID(ref), nil
-}
-
-func (ref EventRef) MarshalJSON() ([]byte, error) {
-	obj, found := DB.GetEventByID(primitive.ObjectID(ref))
-	if !found {
-		return nil, fmt.Errorf("could not find event with id %s", ref)
-	}
-
-	return json.Marshal(obj)
-}
-
-// Same functionality is ObjectID.UnmarshalJSON except it returns an error if the referenced object doesn't exist
-func (ref *EventRef) UnmarshalJSON(b []byte) error {
-	id := primitive.ObjectID(*ref)
-	if err := id.UnmarshalJSON(b); err != nil {
-		return err
-	}
-	_, found := DB.GetEventByID(id)
-	if !found {
-		return fmt.Errorf("object not found")
-	}
-
-	*ref = EventRef(id)
-	return nil
-}
-
-// Gets the referenced object and panics if it doesn't exist
-func (ref EventRef) Get() Event {
-	obj, found := DB.GetEventByID(primitive.ObjectID(ref))
-	if !found {
-		panic("could not find object")
-	}
-	return obj
-}
-
-// Ref creates a reference to the object
-func (obj Event) Ref() EventRef {
-	return EventRef(obj.ID)
-}
-
-// CreateEvent creates an event and adds it to the database. The
-// ID element of the Event will be set if it is successful
-func (db *Database) CreateEvent(event *Event) {
-	result, err := db.Collections.Events.InsertOne(context.TODO(), event)
-	if err != nil {
-		panic(err)
-	}
-
-	event.ID = result.InsertedID.(primitive.ObjectID)
-}
-
-// GetEvents returns a list of all events stored in the database.
-// It will return an ordered list with the most recent first.
-func (db *Database) GetEvents() []Event {
-	cur, err := db.Collections.Events.Find(context.TODO(), bson.D{}, &options.FindOptions{
-		// Sort by date
-		Sort: bson.D{{"time", -1}},
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	events := make([]Event, 0)
-	if err := cur.All(context.TODO(), &events); err != nil {
-		panic(err)
-	}
-
-	return events
-}
-
-// GetLocationByID gets a location by its ID. If not found, found will be false
-func (db *Database) GetEventByID(id primitive.ObjectID) (event Event, found bool) {
-	result := db.Collections.Events.FindOne(nil, bson.M{"_id": id})
-
-	err := result.Err()
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return Event{}, false
-		}
-		panic(err)
-	}
-
-	if err := result.Decode(&event); err != nil {
-		panic(err)
-	}
-
-	found = true
-	return
-}
 
 // GetMostRecentEvent gets the most recent event created by the specified studentID
 // If the event is found, it will be returned, otherwise, found will be false.
