+++ pkg/database/gen-event.go
@@ -1,0 +1,164 @@
+// This file was automatically generated by genny.
+// Any changes will be lost if this file is regenerated.
+// see https://github.com/cheekybits/genny
+
+// This file contains generic code for implementing basic methods
+// for each event such as references, Get by ID, Update, etc...
+// If you're not modifying these functions, you shouldn't have to worry
+// about regenerating code. However, if you updated this file, to update
+// the changes for each of the events you would have to install
+// genny https://github.com/cheekybits/genny and run go generate.
+
+package database
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+// EventRef is a reference to a Event which, when serialized, will return
+// the json of the referenced object.
+//
+// Be careful for circular references.
+type EventRef primitive.ObjectID
+
+func (ref EventRef) GetBSON() (interface{}, error) {
+	return primitive.ObjectID(ref), nil
+}
+
+func (ref EventRef) MarshalJSON() ([]byte, error) {
+	obj, found := DB.GetEventByID(primitive.ObjectID(ref))
+	if !found {
+		return nil, fmt.Errorf("could not find Event with id %s", ref)
+	}
+
+	return json.Marshal(obj)
+}
+
+// Same functionality is ObjectID.UnmarshalJSON except it returns an error if the referenced object doesn't exist
+func (ref *EventRef) UnmarshalJSON(b []byte) error {
+	id := primitive.ObjectID(*ref)
+	if err := id.UnmarshalJSON(b); err != nil {
+		return err
+	}
+	_, found := DB.GetEventByID(id)
+	if !found {
+		return fmt.Errorf("object not found")
+	}
+
+	*ref = EventRef(id)
+	return nil
+}
+
+// Gets the referenced object and panics if it doesn't exist
+func (ref EventRef) Get() Event {
+	obj, found := DB.GetEventByID(primitive.ObjectID(ref))
+	if !found {
+		panic("could not find object")
+	}
+	return obj
+}
+
+// Ref creates a reference to the object
+func (obj Event) Ref() EventRef {
+	return EventRef(obj.ID)
+}
+
+// CreateEvent creates a Event and adds it to the database. The
+// ID element of the newly created Event will be set if it is successful
+func (db *Database) CreateEvent(event *Event) {
+	result, err := db.Collections.Events.InsertOne(context.TODO(), event)
+	if err != nil {
+		panic(err)
+	}
+
+	event.ID = result.InsertedID.(primitive.ObjectID)
+}
+
+// GetEvents returns a list of all events stored in the database.
+func (db *Database) GetEvents() []Event {
+	cur, err := db.Collections.Events.Find(context.TODO(), bson.D{})
+	if err != nil {
+		panic(err)
+	}
+
+	events := make([]Event, 0)
+	if err := cur.All(context.TODO(), &events); err != nil {
+		panic(err)
+	}
+
+	return events
+}
+
+// GetEventByID gets a event by their ID. If not found, found will be false and
+// err will be nil.
+func (db *Database) GetEventByID(id primitive.ObjectID) (event Event, found bool) {
+	result := db.Collections.Events.FindOne(nil, bson.M{"_id": id})
+
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return Event{}, false
+		}
+		panic(err)
+	}
+
+	if err := result.Decode(&event); err != nil {
+		panic(err)
+	}
+
+	found = true
+	return
+}
+
+// GetEventByIDString gets a event by its ID as a string. If the ID could not be
+// parsed into an object ID or the event could not be found, an error will be returned
+func (db *Database) GetEventByIDString(id string) (Event, error) {
+	objectID, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
+		return Event{}, err
+	}
+
+	event, found := db.GetEventByID(objectID)
+	if !found {
+		return Event{}, fmt.Errorf("no Events found with id %s", objectID.Hex())
+	}
+
+	return event, nil
+}
+
+// DeleteEvent deletes a event from the database by ID. If the event could not be
+// found, success will be false
+func (db *Database) DeleteEvent(id primitive.ObjectID) bool {
+	result := db.Collections.Events.FindOneAndDelete(nil, bson.M{"_id": id})
+
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return false
+		}
+		panic(err)
+	}
+	return true
+}
+
+// UpdateEvent finds a event by its ID and updates it
+func (db *Database) UpdateEvent(id primitive.ObjectID, newEvent *Event) bool {
+	result := db.Collections.Locations.FindOneAndUpdate(nil, bson.M{"_id": id}, bson.M{"$set": newEvent})
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return false
+		}
+		panic(err)
+	}
+
+	err = result.Decode(newEvent)
+
+	return true
+}
