+++ pkg/database/gen-location.go
@@ -1,0 +1,164 @@
+// This file was automatically generated by genny.
+// Any changes will be lost if this file is regenerated.
+// see https://github.com/cheekybits/genny
+
+// This file contains generic code for implementing basic methods
+// for each location such as references, Get by ID, Update, etc...
+// If you're not modifying these functions, you shouldn't have to worry
+// about regenerating code. However, if you updated this file, to update
+// the changes for each of the locations you would have to install
+// genny https://github.com/cheekybits/genny and run go generate.
+
+package database
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+// LocationRef is a reference to a Location which, when serialized, will return
+// the json of the referenced object.
+//
+// Be careful for circular references.
+type LocationRef primitive.ObjectID
+
+func (ref LocationRef) GetBSON() (interface{}, error) {
+	return primitive.ObjectID(ref), nil
+}
+
+func (ref LocationRef) MarshalJSON() ([]byte, error) {
+	obj, found := DB.GetLocationByID(primitive.ObjectID(ref))
+	if !found {
+		return nil, fmt.Errorf("could not find Location with id %s", ref)
+	}
+
+	return json.Marshal(obj)
+}
+
+// Same functionality is ObjectID.UnmarshalJSON except it returns an error if the referenced object doesn't exist
+func (ref *LocationRef) UnmarshalJSON(b []byte) error {
+	id := primitive.ObjectID(*ref)
+	if err := id.UnmarshalJSON(b); err != nil {
+		return err
+	}
+	_, found := DB.GetLocationByID(id)
+	if !found {
+		return fmt.Errorf("object not found")
+	}
+
+	*ref = LocationRef(id)
+	return nil
+}
+
+// Gets the referenced object and panics if it doesn't exist
+func (ref LocationRef) Get() Location {
+	obj, found := DB.GetLocationByID(primitive.ObjectID(ref))
+	if !found {
+		panic("could not find object")
+	}
+	return obj
+}
+
+// Ref creates a reference to the object
+func (obj Location) Ref() LocationRef {
+	return LocationRef(obj.ID)
+}
+
+// CreateLocation creates a Location and adds it to the database. The
+// ID element of the newly created Location will be set if it is successful
+func (db *Database) CreateLocation(location *Location) {
+	result, err := db.Collections.Locations.InsertOne(context.TODO(), location)
+	if err != nil {
+		panic(err)
+	}
+
+	location.ID = result.InsertedID.(primitive.ObjectID)
+}
+
+// GetLocations returns a list of all locations stored in the database.
+func (db *Database) GetLocations() []Location {
+	cur, err := db.Collections.Locations.Find(context.TODO(), bson.D{})
+	if err != nil {
+		panic(err)
+	}
+
+	locations := make([]Location, 0)
+	if err := cur.All(context.TODO(), &locations); err != nil {
+		panic(err)
+	}
+
+	return locations
+}
+
+// GetLocationByID gets a location by their ID. If not found, found will be false and
+// err will be nil.
+func (db *Database) GetLocationByID(id primitive.ObjectID) (location Location, found bool) {
+	result := db.Collections.Locations.FindOne(nil, bson.M{"_id": id})
+
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return Location{}, false
+		}
+		panic(err)
+	}
+
+	if err := result.Decode(&location); err != nil {
+		panic(err)
+	}
+
+	found = true
+	return
+}
+
+// GetLocationByIDString gets a location by its ID as a string. If the ID could not be
+// parsed into an object ID or the location could not be found, an error will be returned
+func (db *Database) GetLocationByIDString(id string) (Location, error) {
+	objectID, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
+		return Location{}, err
+	}
+
+	location, found := db.GetLocationByID(objectID)
+	if !found {
+		return Location{}, fmt.Errorf("no Locations found with id %s", objectID.Hex())
+	}
+
+	return location, nil
+}
+
+// DeleteLocation deletes a location from the database by ID. If the location could not be
+// found, success will be false
+func (db *Database) DeleteLocation(id primitive.ObjectID) bool {
+	result := db.Collections.Locations.FindOneAndDelete(nil, bson.M{"_id": id})
+
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return false
+		}
+		panic(err)
+	}
+	return true
+}
+
+// UpdateLocation finds a location by its ID and updates it
+func (db *Database) UpdateLocation(id primitive.ObjectID, newLocation *Location) bool {
+	result := db.Collections.Locations.FindOneAndUpdate(nil, bson.M{"_id": id}, bson.M{"$set": newLocation})
+	err := result.Err()
+	if err != nil {
+		if err == mongo.ErrNoDocuments {
+			return false
+		}
+		panic(err)
+	}
+
+	err = result.Decode(newLocation)
+
+	return true
+}
